requirements.txt
websocket
websocket-client
json
python-binance
datetime
